compilation under ubuntu
added setRegistration to capture in kinfu module.

git-svn-id: svn+ssh://svn.pointclouds.org/pcl/trunk@3233 a9d63959-f2ad-4865-b262-bf0e56cfafb6

+++ cuda/kinfu/src/cuda/extract.cu
@@ -250,7 +250,7 @@ namespace pcl
                     //last block
                     if (value == total_blocks - 1)
                     {                        
-                        output_count = min(output.size, global_count);
+                        output_count = min((int)output.size, global_count);
                         blocks_done = 0;
                         global_count = 0;  
                     }
